Add more MPC message types

+++ PokerGameApp/C++Classes/GameProcedure_CPPWrapper.h
@@ -17,9 +17,9 @@
 - (instancetype)GameProcedure_CPPWrapper;
 - (void)ShuffleCards:(NSMutableArray*)pca1 pca2:(NSMutableArray*)pca2 pca3:(NSMutableArray*)pca3 pca4:(NSMutableArray*)pca4;
 
--(NSArray<Card_CPPWrapper*>*)testStarter:(NSArray<Card_CPPWrapper*>*)cards suit:(int)suit n:(int)n;
+-(NSArray<Card_CPPWrapper*>*)testStarter:(NSArray<Card_CPPWrapper*>*)cards suit:(NSInteger)suit n:(NSInteger)n;
 
-- (bool)remove: (NSArray<Card_CPPWrapper*>*)removeList n:(int)n;
+- (bool)remove: (NSArray<Card_CPPWrapper*>*)removeList n:(NSInteger)n;
 
 @end
 #endif /* GameProcedure_CPPWrapper_h */
